RELEASE_1_2_10 to MASTER

See merge request talosweb/analyst-console-escalations!1778

+++ app/models/dispute.rb
@@ -451,6 +451,8 @@ class Dispute < ApplicationRecord
     user = User.where(cvs_username:"vrtincom").first
     begin
 
+      entry_claims = {}
+
       guest = Company.where(:name => "Guest").first
       opened_at = Time.now
       resolved_at = Time.now
@@ -558,6 +560,14 @@ class Dispute < ApplicationRecord
 
         new_entries_ips.each do |ip, entry|
 
+
+          #if ["Suspicious sites", "High risk","Poor"].include?(entry[:sbrs]["rep_sugg"])
+          #  false_negative_claim = true
+          #end
+          #
+          claim = entry[:sbrs]["claim"]
+          entry_claims[ip] = claim
+
           if entry[:sbrs]["platform"].present?
             entry_platform = Platform.find(entry[:sbrs]["platform"].to_i) rescue nil
           end
@@ -571,6 +581,7 @@ class Dispute < ApplicationRecord
           new_dispute_entry.status = DisputeEntry::NEW
           new_dispute_entry.resolution = ""
           new_dispute_entry.suggested_disposition = entry[:sbrs]["rep_sugg"]
+          new_dispute_entry.suggested_threat_category = entry[:sbrs]["suggested_threat_category"] unless entry[:sbrs]["suggested_threat_category"].blank?
           new_dispute_entry.case_closed_at = resolved_at
           new_dispute_entry.case_resolved_at = resolved_at
 
@@ -580,7 +591,6 @@ class Dispute < ApplicationRecord
           new_dispute_entry.platform_id = entry_platform.id unless entry_platform.blank?
           new_dispute_entry.platform = entry[:sbrs]["platform"] if (entry[:sbrs]["platform"].present? && !entry[:sbrs]["platform"].kind_of?(Integer))
           new_dispute_entry.save!
-
 
           if entry && entry[:wbrs] && entry[:wbrs]["WBRS_Rule_Hits"]
             wbrs_hits = entry[:wbrs]["WBRS_Rule_Hits"].split(",").map {|hit| hit.strip }
@@ -597,6 +607,7 @@ class Dispute < ApplicationRecord
           end
 
           #total_hits = (wbrs_hits + sbrs_hits).uniq
+
 
           sbrs_hits.each do |rule_hit|
             new_rule_hit = DisputeRuleHit.new
@@ -616,9 +627,10 @@ class Dispute < ApplicationRecord
 
         end
 
-
-
         new_entries_urls.each do |url, entry|
+
+          claim = entry["claim"]
+          entry_claims[url] = claim
 
           if entry["platform"].present?
             entry_platform = Platform.find(entry["platform"].to_i) rescue nil
@@ -633,7 +645,7 @@ class Dispute < ApplicationRecord
           new_dispute_entry.resolution = ""
           new_dispute_entry.case_closed_at = resolved_at
           new_dispute_entry.case_resolved_at = resolved_at
-
+          new_dispute_entry.suggested_threat_category = entry["suggested_threat_category"] unless entry["suggested_threat_category"].blank?
           new_dispute_entry.case_opened_at = opened_at
           new_dispute_entry.wbrs_score = entry["WBRS_SCORE"] == "No score" ? nil : entry["WBRS_SCORE"]
 
@@ -678,7 +690,6 @@ class Dispute < ApplicationRecord
 
       end
 
-
       ######record creation completed######
 
       ######AUTO RESOLVE LOGIC########
@@ -698,9 +709,7 @@ class Dispute < ApplicationRecord
 
           ########Auto Resolve for IP addressses (email)##############
           if dispute_entry.entry_type == "IP"
-            if ["Suspicious sites", "High risk","Poor"].include?(dispute_entry.suggested_disposition)
-              false_negative_claim = true
-            end
+            entry_claim = entry_claims[dispute_entry.hostlookup]
 
             logger.info "fetching preload"
 
@@ -713,7 +722,7 @@ class Dispute < ApplicationRecord
 
             if !matching_disposition
 
-              if !false_negative_claim
+              if entry_claim != "false negative"
                 dispute_entry.status = DisputeEntry::NEW
 
                 if new_dispute.submitter_type == "NON-CUSTOMER" && new_dispute.submission_type == "e"
@@ -733,13 +742,10 @@ class Dispute < ApplicationRecord
 
           end
 
-
           ############################################################
           #########Auto Resolve for URLs (web)########################
           if dispute_entry.entry_type == "URI/DOMAIN"
-            if ["Suspicious sites", "High risk","Poor"].include?(dispute_entry.suggested_disposition)
-              false_negative_claim = true
-            end
+            entry_claim = entry_claims[dispute_entry.hostlookup]
 
             logger.info "fetching preload"
 
@@ -760,7 +766,7 @@ class Dispute < ApplicationRecord
             end
 
             begin
-              dispute_entry.is_important = is_important?(key)
+              dispute_entry.is_important = is_important?(dispute_entry.hostlookup)
             rescue => e
               Rails.logger.error e
               Rails.logger.error e.backtrace.join("\n")
@@ -770,7 +776,7 @@ class Dispute < ApplicationRecord
 
             if !matching_disposition
 
-              if !false_negative_claim
+              if entry_claim != "false negative"
                 dispute_entry.update(status: DisputeEntry::NEW)
               else
                 dispute_entry = AutoResolve.attempt_ai_conviction(dispute_entry.dispute_rule_hits.pluck(:name), dispute_entry)
@@ -817,13 +823,17 @@ class Dispute < ApplicationRecord
       Rails.logger.error "Dispute failed to save, backing out all DB changes."
       Rails.logger.error $!
       Rails.logger.error $!.backtrace.join("\n")
-      new_dispute.reload
-      new_dispute.dispute_entries.destroy_all
-      new_dispute.destroy
-      conn = ::Bridge::DisputeFailedEvent.new(addressee: "talos-intelligence", source_authority: "talos-intelligence", source_key: message_payload["source_key"])
-      conn.post
+      if new_dispute.present?
+        new_dispute.reload
+        new_dispute.dispute_entries.destroy_all
+        new_dispute.destroy
+      end
+      if message_payload["source_key"].present?
+        conn = ::Bridge::DisputeFailedEvent.new(addressee: "talos-intelligence", source_authority: "talos-intelligence", source_key: message_payload["source_key"])
+        conn.post
+      end
 
-      nil
+      return nil
     end
 
     new_dispute.reload
@@ -832,7 +842,7 @@ class Dispute < ApplicationRecord
       new_dispute.destroy
       conn = ::Bridge::DisputeFailedEvent.new(addressee: "talos-intelligence", source_authority: "talos-intelligence", source_key: message_payload["source_key"])
       conn.post
-      nil
+      return nil
     end
 
     new_dispute
@@ -1339,7 +1349,10 @@ class Dispute < ApplicationRecord
 
     disputes_ary = []
     Dispute.transaction do
-      disputes = Dispute.where(id: dispute_ids, status: [Dispute::STATUS_NEW, Dispute::STATUS_REOPENED])
+      disputes = Dispute.where(id: dispute_ids).where.not(status: [
+        Dispute::TI_NEW, Dispute::STATUS_RESOLVED, Dispute::STATUS_RESOLVED_FIXED_FP, Dispute::STATUS_RESOLVED_FIXED_FN,
+        Dispute::STATUS_RESOLVED_UNCHANGED
+      ])
       disputes_ary = disputes.all.to_a
       disputes.update_all(user_id: user_id, status: Dispute::STATUS_ASSIGNED, case_accepted_at: accepted_at)
